+++ app/forked/defaults.rb
@@ -1,0 +1,11 @@
+module Forked
+  extend self
+
+  def forked_defaults
+    {
+      # if autosave is true, game progress will be
+      # saved every time the player clicks a button
+      autosave: true,
+    }
+  end
+end