+++ app/main.rb
@@ -1,4 +1,6 @@
 require 'app/libs/json.rb'
+
+require 'app/forked/defaults.rb'
 require 'app/forked/tools.rb'
 require 'app/forked/themes/light-mode-theme.rb'
 require 'app/forked/themes/dark-mode-theme.rb'